+++ test/src/client_utils.dart
@@ -18,6 +18,7 @@ import 'dart:convert';
 
 import 'package:grpc/grpc.dart';
 import 'package:grpc/src/client/channel.dart' as base;
+import 'package:grpc/src/client/client_keepalive.dart';
 import 'package:grpc/src/client/http2_connection.dart';
 import 'package:grpc/src/shared/message.dart';
 import 'package:http2/transport.dart';
@@ -76,6 +77,9 @@ class FakeChannelOptions implements ChannelOptions {
   BackoffStrategy backoffStrategy = testBackoff;
   @override
   CodecRegistry codecRegistry = CodecRegistry.empty();
+
+  @override
+  ClientKeepAliveOptions get keepAlive => const ClientKeepAliveOptions();
 }
 
 class FakeChannel extends ClientChannel {