+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 3.2.3-wip
+
+* Add const constructor to `GrpcError` fixing #606.
+* Make `GrpcError` non-final to allow implementations.
+* Only send keepalive pings on open connections.
+
+## 3.2.2
+
+* Remove `base` qualifier on `ResponseStream`.
+* Add support for clients to send KEEPALIVE pings.
+
+## 3.2.1
+
+* `package:http` now supports more versions: `>=0.13.0 <2.0.0`. 
+* `package:protobuf` new supports more versions: `>=2.0.0 <4.0.0`.
+
 ## 3.2.0
 
 * `ChannelOptions` now exposes `connectTimeout`, which is used on the